requirements.txt
# NEOwatch dependencies
python-telegram-bot[job-queue]==21.10
requests
fuzzywuzzy
python-Levenshtein
python-dotenv
